TexturePackerのリファクタ：ファイル訳
